+++ BaoMiWanBiao/SleepStatusViewController.m
@@ -191,7 +191,7 @@
 {
     NSArray *dateArr = [self.fmTool queryData:dateStr];
     
-    NSLog(@"%ld",dateArr.count);
+    NSLog(@"%ld",(unsigned long)dateArr.count);
     if (dateArr.count != 0 ) {
         self.SleepModel = dateArr.firstObject;
